Отключение лишних компонентов в cli-режиме
