+++ src/main/java/ru/digilabs/alkir/rahc/auth/AuthConfigurerAdapter.java
@@ -3,6 +3,7 @@ package ru.digilabs.alkir.rahc.auth;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 import org.springframework.core.annotation.Order;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
@@ -14,6 +15,7 @@ import ru.digilabs.alkir.rahc.configuration.AuthConfigurationProperties;
 
 @Configuration
 @RequiredArgsConstructor
+@Profile("web")
 @EnableWebSecurity
 @Order(1)
 public class AuthConfigurerAdapter {
