+++ src/main/resources/application-cli.properties
@@ -1,3 +1,20 @@
 spring.main.banner-mode=off
 logging.level.ru.digilabs.alkir.rahc.RahcApplication=off
 logging.level.com._1c.v8.swp.internal.netty.Connector=off
+
+spring.application.admin.enabled=false
+management.server.port=-1
+management.endpoints.jmx.exposure.exclude=*
+management.endpoints.web.exposure.exclude=*
+management.auditevents.enabled=false
+management.health.defaults.enabled=false
+management.observations.enable.all=false
+management.metrics.enable.all=false
+management.defaults.metrics.export.enabled=false
+management.info.defaults.enabled=false
+spring.jmx.enabled=false
+
+spring.sql.init.mode=never
+spring.thymeleaf.enabled=false
+
+spring.autoconfigure.exclude+=org.springframework.boot.actuate.autoconfigure.observation.ObservationAutoConfiguration,org.springframework.boot.actuate.autoconfigure.metrics.JvmMetricsAutoConfiguration,org.springframework.boot.autoconfigure.jmx.JmxAutoConfiguration,org.springframework.boot.actuate.autoconfigure.endpoint.jmx.JmxEndpointAutoConfiguration,org.springframework.boot.autoconfigure.http.client.HttpClientAutoConfiguration,org.springframework.boot.autoconfigure.web.client.RestClientAutoConfiguration,org.springframework.boot.autoconfigure.web.reactive.function.client.WebClientAutoConfiguration,,org.springframework.boot.actuate.autoconfigure.metrics.SystemMetricsAutoConfiguration,org.springframework.boot.actuate.autoconfigure.observation.web.client.HttpClientObservationsAutoConfiguration,org.springframework.boot.autoconfigure.web.reactive.function.client.ClientHttpConnectorAutoConfiguration,org.springframework.boot.autoconfigure.http.codec.CodecsAutoConfiguration,org.springframework.boot.actuate.autoconfigure.endpoint.EndpointAutoConfiguration,org.springframework.boot.actuate.autoconfigure.endpoint.web.WebEndpointAutoConfiguration,org.springframework.boot.actuate.autoconfigure.metrics.CompositeMeterRegistryAutoConfiguration,org.springframework.boot.autoconfigure.http.HttpMessageConvertersAutoConfiguration,org.springframework.boot.autoconfigure.http.client.HttpClientAutoConfiguration,org.springframework.boot.actuate.autoconfigure.metrics.MetricsAutoConfiguration
